Package refactoring + outbox mechanism implemented + work manager initial setup

+++ app/src/main/java/com/mohsinsyed/aac_sample/data/remote/PostService.kt
@@ -1,6 +1,6 @@
 package com.mohsinsyed.aac_sample.data.remote
 
-import com.mohsinsyed.aac_sample.data.models.Post
+import com.mohsinsyed.aac_sample.data.models.entities.Post
 import com.mohsinsyed.aac_sample.utils.constants.AppConstants
 import retrofit2.http.*
 
@@ -9,10 +9,10 @@ interface PostService {
     suspend fun create(@Body post: Post?): Post?
 
     @PUT(AppConstants.NetworkConstants.MODIFY_POST)
-    suspend fun update(@Body post: Post?, @Path("postId") id: Int?): Post?
+    suspend fun update(@Body post: Post?, @Path("postId") id: Long?): Post?
 
     @DELETE(AppConstants.NetworkConstants.MODIFY_POST)
-    suspend fun delete(@Path("postId") id: Int?): Unit?
+    suspend fun delete(@Path("postId") id: Long?): Unit?
 
     @GET(AppConstants.NetworkConstants.FETCH_POSTS_BY_USER_ID)
     suspend fun fetchAll() : List<Post>?
